commit : 메인 페이지 데이터만 컴포넌트로 분리
